+++ core/__init__.py
@@ -1,3 +1,3 @@
 """Focomy - The Most Beautiful CMS"""
 
-__version__ = "0.1.41"
+__version__ = "0.1.42"